backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.2
opencv-python==4.8.0.76
scikit-learn==1.3.0
tqdm==4.65.0
scipy==1.11.1
seaborn==0.12.2 